requirements.txt
# Core Web Framework
fastapi==0.129.0
uvicorn==0.41.0
gunicorn==23.0.0
pydantic==2.12.5

# Database
motor==3.7.1
pymongo==4.16.0
dnspython==2.8.0

# Authentication & Security
python-jose==3.5.0
bcrypt==5.0.0
argon2-cffi==25.1.0
passlib==1.7.4
email-validator==2.3.0

# Environment & Configuration
python-dotenv==1.2.1
python-dateutil==2.9.0.post0

# HTTP & API Clients
requests==2.32.5
httpx==0.27.0

# Payment Gateway
razorpay==2.0.0

# Rate Limiting
slowapi==0.1.9

# Task Scheduling
apscheduler==3.11.2
